RDKAAMP-1380: Latency Correction Issue in LLD
Reason for change: Added common logic to LLD and non LLD
for handling rate correction; Which resolve the 0.9 to 1 transformation
Issue from in LLD
Test Procedure: Play new LLD channel and moniter rate changes

Risks: Low


RDKAAMP-1380: Latency Correction Issue in LLD

Reason for change: Follow up change to correct the minimum buffer needed
for rate change
Test Procedure: Play new LLD channel and moniter latency;
No freeze due to 0 buffer is not expected due to rate corretion

Risks: Low

+++ priv_aamp.cpp
@@ -2045,12 +2045,12 @@ void PrivateInstanceAAMP::ReportProgress(bool sync, bool beginningOfStream)
 				static int tick;
 				if ((tick++ % 4) == 0)
 				{
-					AAMPLOG_WARN("aamp pos: [%ld..%ld..%ld..%lld..%ld..%.2f..%s..%ld..%ld]",
+					AAMPLOG_WARN("aamp pos: [%ld..%ld..%ld..%lld..%.2f..%.2f..%s..%ld..%ld]",
 						(long)(start / 1000),
 						(long)(reportFormatPosition / 1000),
 						(long)(end / 1000),
 						(long long) videoPTS,
-						(long)(bufferedDuration / 1000),
+						(double)(bufferedDuration / 1000.0),
 					        (latency / 1000),
 						seiTimecode.c_str(),
 						mpStreamAbstractionAAMP->GetVideoBitrate(),
